- 펫 디테일 작업 중

+++ app/src/main/java/xlab/world/xlab/view/topicDetail/TopicPetDetailActivity.kt
@@ -268,7 +268,7 @@ class TopicPetDetailActivity : AppCompatActivity(), View.OnClickListener {
                     RunActivity.petEditActivity(context = this, petNo = null)
                 }
                 R.id.petEditBtn -> { // 펫 수정하기
-                    RunActivity.petEditActivity(context = this, petNo = textViewPetNo.text.toString().toInt())
+                    RunActivity.petEditActivity(context = this, petNo = textViewPetNo.text.toString())
                 }
             }
         }
